\name{type_const}
\alias{type_const}
\docType{data}

\title{
A set of server constants stored in a list
}

\description{
This is a global variable of type list that contains a set of constants used in the Jazz API. All constants are integers. Direct usage of the integer values is possible but not recommended.
}

\examples{
\dontrun{
type_const[['BLOCKTYPE_RAW_MIME_HTML']]
}
}
